PR feedback: refactor schema validation context to contain shared code
